Update and rename clone-nbh.py to ssr_api_lib.py
split code in library and script
